Allow `--project=@scriptdir` for easier co-packaging of scripts and p… (#50864)
…rojects

I have a directory with a number of runnable scripts and a Project.toml
that applies to all of them. I was hoping to have a simple shebang line
for these, but there didn't really seem to be a good way to set the
`--project` without forcing the script to be invoked with a specific cwd
or playing complicated shebang games. With this PR, you can simply
write:

```
#!/usr/bin/env -S julia --project=@scriptdir/..
```

and everything will work out of the box. Perhaps we already have a good
mechanism for this, but if so, I was unable to find it, so in that case,
consider this a documentation issue instead ;).