Use rr-safe `nopl; rdtsc` sequence
When running under `rr`, it needs to patch out `rdtsc` to record the values returned.
If this is not possible, `rr` falls back to an expensive signal-based emulation.
As of rr master, a specific `nopl; rdtsc` sequence may be used to guarantee that
`rdtsc` patching is always possible. Use this sequence for uses of rdtsc in our
runtime.