Hide message about changed ENV when test is ignored (#50883)
PR #47148 added the ability to ignore changes to the environment in a
testsuite (using the `JULIA_TEST_CHECK_MUTATED_ENV` environment
variable), but didn't hide the printing of the error message. This hides
the actual error message so that the terminal isn't cluttered with them
when they should be ignored. This is needed because in the MKL.jl
testsuite, the error is hit for basically every linear algebra test
(https://github.com/JuliaLinearAlgebra/MKL.jl/issues/130), and so even
when we have already disabled the error, printing the error message
causes lots of noise on the terminal that can possibly hide other
problems.

Co-authored-by: Dilum Aluthge <dilum@aluthge.com>