inference: Wrap `unreachable` stmts in `Const(...)`
We were doing this before, but the previous commit had changed it. The
problem is that the optimizer cannot rely on the `unreachable` vector
for its external entrypoint.

This restores the existing logic so that type inference will give you a
result that you can pipe into the optimizer and have it work, although
suboptimally since we won't be able to prune any unreachable branches.