inference: fix return_type_tfunc modeling of concrete functions (#51042)
The `aft` parameter is a value already, so we should be checking it in
the value domain, not the type domain like `tt`. That check happens to
already be done (somewhat unnecessarily) earlier in the function.

Fixes #40606

---------

Co-authored-by: Shuhei Kadowaki <40514306+aviatesk@users.noreply.github.com>
(cherry picked from commit f24a93a3f6607f569c739cbcd3a84d21bdc6c908)