Fix comment in `load_path_expand()` (#50934)
When the meaning of `"@"` in `LOAD_PATH` was changed and `"@."`
introduced as the new name for the previous meaning of `"@"` in
9a4ecaa4e4, a comment in `load_path_expand()` wasn’t updated to reflect
this change. This left the comment being incorrect, leading to potential
confusion.

I hope I’m not creating more hassle than it’s worth with this super-tiny
PR, but since it’s caused some confusion (e. g.
https://github.com/JuliaLang/Pkg.jl/pull/3547#issuecomment-16795025350),
I thought it might be worthwhile to fix it.