cglobal: Add an option to bypass jlplt mechanism
This adds an extra argument to `cglobal` to opt out of the jlplt
runtime symbol resolution mechanism. This is necessary for
StaticCompiler-like code that needs to function in environments
that do not provide the julia runtime symbol resolution support
code. Symbol resolution is deferred to the host linker (static
or dynamic, where the latter may of course use its own PLT).