Workaround upstream FreeBSD issue #272992
FreeBSD does not support looking up WEAK symbols in its libc via `dlsym`
if you're using a library-specific handle. It will give you the empty
stub in libc instead of, e.g., the strong implementation in ld-elf.so

This fix and the one that was merged upstream are pretty conspicuously
incomplete, since `dlsym`, `dladdr` and a ton of other symbols still
have stubs with no trampoline in libc:
https://cgit.freebsd.org/src/tree/lib/libc/gen/dlfcn.c?id=21a52f99440c9bec7679f3b0c5c9d888901c3694

Thankfully, this appears to be the only function that we directly
`ccall` from Julia's Libdl so we can leave this fix incomplete for now.