Add GC write barrier in jl_reserve_excstack (#51096)
`jl_push_excstack` and `jl_reserve_excstack` take an address to the
field `excstack` in `jl_task_t`, and may update the field. In such a
case, a GC write barrier is needed. This PR adds the missing write
barrier.

Note: I am not sure if this affects the correctness of Julia GC or not.
This could be special cased in the GC code, as the task is always
rooted. However, generally if the task is in the old generation, it does
not have to be scanned even if it is a root. A write barrier is needed
to make sure the GC is aware of the update.