Consistently report sysimg size limits
We used to be reporting the maximum system image size in hexadecimal, while reporting the image that was just built in decimal.  This is a recipe for confusion when the size limit contains no blatantly hexadecimal characters.

This PR addresses the issue by consistently printing out both values in hexadecimal (making this more obvious by prefixing the numbers with `0x`).  In order to use `PRIxPTR` in both format strings, I have cast the current image size to `uintptr_t` from `intmax_t`, which should be safe in all situations.