sysimg: Allow loading a system image that is already present in memory
I've written this code probably three times at this point, but for some
reason it never made it into a PR. This allows loading a system image
that has already been loaded into memory. This happen when wanting to
distribute a static or mostly-static binary of julia code where the
system image (and optionally other libraries like libjulia, etc.) are
linked directly into the main executable. It is also useful for deployment
to environments that do not have (or have incomplete) support for dynamic
linking (e.g. wasm or embedded).