Try to fix the wine/mingw build
For some reason I'm getting
```
/home/keno/julia-win64/cli/loader_lib.c:63: undefined reference to `memset'
```
Trying to build for win64 under wine. I'm assuming the memset symbol is being inserted by the compiler for the big zero initialization of the array (perhaps it is usually a preprocessor define). Linking `msvcrt` made this work for me, but I'm not sure if this was something we had intentionally omitted.