Add options to jl_dump_native (#51124)
I am attempting to more or less unify the codepaths between
sysimage-like binary emissions and StaticCompiler-like binary emissions
(the latter currently uses a GPUCompiler-derived pipeline), with a view
towards enabling things that are somewhat in the middle (may have parts,
but not all of the runtime, etc.). To that end, add a parameters struct
to jl_dump_native that will likely grow over time, but for now, the
primary thing it controls is whether or not to emit the sysimage
"metadata" section.