Document and clean up CodegenParams
- Add documentation to CodegenParams fields per request in #51123
- Fix compare_cgparams which hadn't been updated for recent additions
- Remove unused and untested generic_context

The latter was a codegen option that I added, but eventually ended
up not using anywhere, so remove it for the time being. That said,
I may end up in a situation where I need it again in the very near
future, so I may end up eating my words here, but if I need to put
it back, I'll include a test at least ;).