Update advice on safe use of finalizers
This commit concentrates the advice in the place that users of
`finalizer` are most likely to see it and provides examples.

It's unclear to me if any of this is required if you use the
`ReentrantLock`, because `_trylock` will disable finalizers anyway, but
I couldn't work it out and I figured this PR might summon someone who
knows definitively.