slot2ssa: Fix spurious φᶜ edges
The unreachable here seems to be caused by the fact that (as of #50299) we
re-use a more narrow type from Inference that correctly ignores these
edges, but when inserting the `φᶜ` node in `slot2reg` we were including
extra edges that never get exercised at runtime.

I'm not sure _why_ this causes us to hit an unreachable, since the narrow
type from inference is technically still valid (the catch block will never
observe these spurious assignments at runtime), but this seems to fix the
issue and anyway improves the quality of the IRCode generated by `slot2ssa`.

Resolves #51159