testdefs: make sure that if a test set changes the active project, they change it back when they're done (#51029)
We already check the following for mutation by a test set:
1. DEPOT_PATH
2. LOAD_PATH
3. ENV

So this PR just adds the active project to the list of things we check.

Changing the active project during a test set can definitely have
negative effects on subsequent test sets that are run on the same
worker, so we want to make sure if a test set changes the active
project, that they change it back when they're done.

(cherry picked from commit 106e867e6c28347e42b91905d526cc4908e50ec3)