slot2ssa: consider liveness when inserting `PiNode`s
PhiNodes are not required to be inserted for blocks where a variable is
dead, so we can end up with "incoming" values to the block that aren't
valid dominators.

This is benign since (un-)liveness means that these values should never
end up in the IR, but if we blindly insert a PiNode, that can create a
usage "out of thin air" that references a non-dominating value and
produces invalid IR.

This commit resolves this by saving the liveness information from
PhiNode insertion and checking against it for PiNode insertion.