Random: allow negative seeds (#51416)
Alternative to #46190, see that PR for background. There isn't a strong
use-case for accepting negative seeds, but probably many people tried
something like `seed = rand(Int); seed!(rng, seed)` and saw it failing.
As it's easy to support, let's do it.

This might "break" some random streams, those for which the upper bit of
`make_seed(seed)[end]` was set, so it's rare.