add variant of regex.jl:match() that updates idx
Allow `idx` to be a `Ref{Int}`, so it can be updated easily to always
point at the first character not yet matched. This is particularly
useful e.g. when writing parsers that repeatedly call `match()` as a token
scanner on the same string, allowing them to start the next match
right after the previous one, usually with the regex anchored to the
`idx` position with `\G`.

fixes #51429