Use a single Float16 ABI.
Currently, we are using different Float16 ABIs depending on which GCC compiler
Julia is compiled with (i.e., not based on what the user's OS uses), because
of how the runtime intrinsics are implemented in C. Before GCC 12, we pass
Float16 as UInt16, and otherwise we use `half` which passes (on X86) in XMM
registers.

This PR moves to always using the "new" ABI, regardless of the GCC compiler,
by returning Float16's from C using the `float` datatype. This probably only
works on X86, and may need additional code paths for other platforms.