clarify docs on for loops for general iterators (#51639)
This PR updates the `for` loop documentation in the manual, in two ways:

1. It moves the information that `for` can loop over arbitrary
containers up, before the more technical documentation about the scoping
rules. I feel like the more basic information should go first.
2. It says a few more words about what a general iterator is, and gives
an example of a `StepRange`. That's in response to [this discourse
thread](https://discourse.julialang.org/t/julia-documentation-about-for-loop/104725)
that non-unitrange loops are important for beginners but are somewhat
buried.