clean up identifiers defined in `Main` (#51411)
Loaded packages do not need explicit bindings, and the name
`MainInclude` does not need to be visible.

I noticed that `Main` tab completes to `MainInclude`, which is not great
since it's an implementation detail. It's also a bit messy that
`InteractiveUtils` appears in `varinfo()`. This is not necessary to
access its functionality.

The only behavior change from this is that adding a method to `include`
or `eval` in Main used to work since these were explicitly imported.
This now gives the "must be explicitly imported to be extended" error,
which I consider a big improvement. Nobody should be doing that anyway.