Add note in constructor doc about infix notation (#51610)
When reading the documentation---and later, when I was doing an Exercism
example implementing the complex numbers---it wasn't clear to me how
certain symbols could work as infix operators. The Julia documentation
uses ⊘ for rational division in its example, but it doesn't say why this
works for `1 ⊘ 2`. I just added a note to clarify this.